requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-dateutil==2.9.0.post0
openai==1.50.2
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
pytest==8.3.2
pytest-mock==3.14.0
